# MonolithApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Monolith api",
    "description": "This is an api of our monolith backend for our Architecture class (M1, Expert Dev Logiciel Mob Iot, YNOV)",
    "contact": {
      "name": "Sada Demba Thiam",
      "url": "https://github.com/SadaDemba"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/ProductReductions/Shop/{shopId}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReduction"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReduction"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductReduction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductReductions/Paginated/Shop/{shopId}": {
      "get": {
        "tags": [
          "ProductReductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Products/Shop/{shopId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reductions/Seller/{userId}/Paginated/Shop/{shopId}": {
      "get": {
        "tags": [
          "Reductions"
        ],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReductionResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReductionResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReductionResponseResource"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Shops/Paginated": {
      "get": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shops/Paginated/Keyword": {
      "get": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shops/ShopName": {
      "get": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shops/Seller/Owner": {
      "get": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopResponseResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shops/{id}": {
      "get": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shops/Seller/{userId}/{id}": {
      "put": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Shops/Seller/{userId}": {
      "post": {
        "tags": [
          "Shops"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Shop"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shop"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Product": {
        "required": [
          "name",
          "price",
          "productTypeId",
          "shopId",
          "stock"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "shopId": {
            "type": "integer",
            "format": "int32"
          },
          "productTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          },
          "shop": {
            "$ref": "#/components/schemas/Shop"
          },
          "productReductions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductReduction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductReduction": {
        "required": [
          "isActivated",
          "productId",
          "reductionId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "reductionId": {
            "type": "integer",
            "format": "int32"
          },
          "isActivated": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reduction": {
            "$ref": "#/components/schemas/Reduction"
          }
        },
        "additionalProperties": false
      },
      "ProductReductionResponseResource": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "firstPage": {
            "type": "string",
            "nullable": true
          },
          "lastPage": {
            "type": "string",
            "nullable": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "string",
            "nullable": true
          },
          "previousPage": {
            "type": "string",
            "nullable": true
          },
          "currentPage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductReduction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Reduction": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "shopId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "percentage": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "beginDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "shop": {
            "$ref": "#/components/schemas/Shop"
          }
        },
        "additionalProperties": false
      },
      "ReductionResponseResource": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "firstPage": {
            "type": "string",
            "nullable": true
          },
          "lastPage": {
            "type": "string",
            "nullable": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "string",
            "nullable": true
          },
          "previousPage": {
            "type": "string",
            "nullable": true
          },
          "currentPage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reduction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Shop": {
        "required": [
          "ownerId",
          "shopName"
        ],
        "type": "object",
        "properties": {
          "shopId": {
            "type": "integer",
            "format": "int32"
          },
          "shopName": {
            "minLength": 1,
            "type": "string"
          },
          "shopDescription": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "minLength": 1,
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ShopResponseResource": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "firstPage": {
            "type": "string",
            "nullable": true
          },
          "lastPage": {
            "type": "string",
            "nullable": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "string",
            "nullable": true
          },
          "previousPage": {
            "type": "string",
            "nullable": true
          },
          "currentPage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Shop"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}